Update to OpenZL release v0.2.3

+++ lz/openzl/src/openzl/compress/cnode.c
@@ -208,7 +208,7 @@ ZL_DictID CNODE_getDictID(CNode const* cnode)
     return cnode->transformDesc.publicDesc.dictID;
 }
 
-size_t CNODE_getDictIndex(CNode const* cnode)
+uint32_t CNODE_getDictIndex(CNode const* cnode)
 {
     ZL_ASSERT_NN(cnode);
     ZL_ASSERT_EQ(cnode->nodetype, node_internalTransform);
